fix: remove circular dependencies from wallet-sdk
