+++ packages/wallet-sdk/src/index.ts
@@ -2,5 +2,6 @@ export * from './generate';
 export * from './derive';
 export * from './models/wallet';
 export * from './models/account';
+export * from './models/common';
 export * from './models/wallet-config';
 export * from './encryption';
