+++ packages/wallet-sdk/src/models/wallet.ts
@@ -1,37 +1,12 @@
-import { fromBase58 } from 'bip32';
 import { deriveAccount, deriveLegacyConfigPrivateKey } from '../derive';
 import { connectToGaiaHubWithConfig, getHubInfo } from '../utils';
-import { Account } from './account';
+import { Wallet, getRootNode } from './common';
 import { fetchLegacyWalletConfig } from './legacy-wallet-config';
 import { fetchWalletConfig, updateWalletConfig, WalletConfig } from './wallet-config';
-
-/**
- * This object represents the keys that were derived from the root-level
- * keychain of a wallet.
- */
-export interface WalletKeys {
-  /** Used when generating app private keys, which encrypt app-specific data */
-  salt: string;
-  /** The private key associated with the root of a BIP39 keychain */
-  rootKey: string;
-  /** A private key used to encrypt configuration data */
-  configPrivateKey: string;
-}
-
-export interface Wallet extends WalletKeys {
-  /** The encrypted secret key */
-  encryptedSecretKey: string;
-  /** A list of accounts generated by this wallet */
-  accounts: Account[];
-}
 
 export interface LockedWallet {
   encryptedSecretKey: string;
 }
-
-export const getRootNode = (wallet: Wallet) => {
-  return fromBase58(wallet.rootKey);
-};
 
 /**
  * Restore wallet accounts by checking for encrypted WalletConfig files,
